+++ common/src/python/error_logging/error_logger.py
@@ -2,7 +2,6 @@ import logging
 from datetime import datetime as dt
 from typing import Any, Dict, List, Literal, Optional
 
-from dates.form_dates import DEFAULT_DATE_TIME_FORMAT
 from flywheel.models.file_entry import FileEntry
 from flywheel.rest import ApiException
 from flywheel_adaptor.flywheel_proxy import ProjectAdaptor
@@ -15,6 +14,7 @@ from nacc_common.data_identification import (
     VisitIdentification,
 )
 from nacc_common.error_models import FileErrorList, FileQCModel, QCStatus
+from nacc_common.form_dates import DEFAULT_DATE_TIME_FORMAT
 from pydantic import BaseModel, ValidationError
 from utils.decorators import api_retry, retry_with_backoff
