+++ common/src/python/preprocess/preprocessor_helpers.py
@@ -7,13 +7,13 @@ from datetime import datetime
 from typing import Any, Dict, List, Optional
 
 from configs.ingest_configs import ModuleConfigs
-from dates.form_dates import build_date
 from nacc_common.data_identification import (
     DataIdentification,
 )
 from nacc_common.field_names import (
     FieldNames,
 )
+from nacc_common.form_dates import build_date
 from outputs.error_writer import ErrorWriter
 from outputs.errors import (
     preprocess_errors,
