+++ common/test/python/enrollment/test_enrollment_serialization.py
@@ -4,17 +4,17 @@ from datetime import datetime
 from typing import Dict
 
 import pytest
-from dates.form_dates import (
-    DATE_FORMATS,
-    DateFormatException,
-    parse_date,
-)
 from enrollment.enrollment_transfer import (
     Demographics,
     EnrollmentRecord,
     TransferRecord,
 )
 from identifiers.model import CenterIdentifiers
+from nacc_common.form_dates import (
+    DATE_FORMATS,
+    DateFormatException,
+    parse_date,
+)
 from pydantic import ValidationError
