+++ gear/legacy_identifier_transfer/src/python/legacy_identifier_transfer_app/main.py
@@ -5,12 +5,12 @@ from datetime import datetime
 from typing import Dict, List, Mapping, Optional
 
 from datastore.forms_store import FormsStore
-from dates.form_dates import DEFAULT_DATE_FORMAT, DateFormatException, parse_date
 from enrollment.enrollment_project import EnrollmentProject
 from enrollment.enrollment_transfer import EnrollmentRecord
 from identifiers.model import CenterIdentifiers, IdentifierObject
 from keys.keys import DefaultValues, MetadataKeys
 from nacc_common.field_names import FieldNames
+from nacc_common.form_dates import DEFAULT_DATE_FORMAT, DateFormatException, parse_date
 from notifications.email import EmailClient, create_ses_client
 from pydantic import ValidationError
 from uploads.upload_error import UploaderError
