+++ gear/participant_transfer/src/python/participant_transfer_app/transfer.py
@@ -5,7 +5,6 @@ from datetime import datetime
 from typing import List, Optional
 
 from centers.center_group import CenterGroup
-from dates.form_dates import DEFAULT_DATE_FORMAT
 from enrollment.enrollment_project import EnrollmentProject, TransferInfo
 from enrollment.enrollment_subject import EnrollmentSubject
 from enrollment.enrollment_transfer import (
@@ -27,6 +26,7 @@ from identifiers.identifiers_repository import IdentifierUpdateObject
 from identifiers.model import CenterIdentifiers, IdentifierObject
 from keys.keys import DefaultValues
 from nacc_common.data_identification import DataIdentification
+from nacc_common.form_dates import DEFAULT_DATE_FORMAT
 from pydantic import ValidationError
 from uploads.upload_error import UploaderError
